Move 'geodata' signals to signals.py
