+++ rgd/geodata/models/fmv/base.py
@@ -2,9 +2,6 @@ import base64
 import pickle
 
 from django.contrib.gis.db import models
-from django.db import transaction
-from django.db.models.signals import post_save
-from django.dispatch import receiver
 from s3_file_field import S3FileField
 
 from rgd.utility import _link_url
@@ -26,11 +23,6 @@ class FMVFile(ChecksumFile, TaskEventMixin):
         return _link_url('geodata', 'fmv_file', self, 'file')
 
     fmv_data_link.allow_tags = True
-
-
-@receiver(post_save, sender=FMVFile)
-def _post_save_fmv_file(sender, instance, *args, **kwargs):
-    transaction.on_commit(lambda: instance._post_save_event_task(*args, **kwargs))
 
 
 class FMVEntry(ModifiableEntry, SpatialEntry):
