+++ rgd/geodata/models/imagery/ifiles.py
@@ -1,8 +1,5 @@
 """Models for handing input files."""
 from django.contrib.gis.db import models
-from django.db import transaction
-from django.db.models.signals import post_save
-from django.dispatch import receiver
 from s3_file_field import S3FileField
 
 from rgd.utility import _link_url
@@ -45,11 +42,6 @@ class ImageFile(ChecksumFile, TaskEventMixin, BaseImageFile):
         return _link_url('geodata', 'image_file', self, 'file')
 
     image_data_link.allow_tags = True
-
-
-@receiver(post_save, sender=ImageFile)
-def _post_save_image_file(sender, instance, *args, **kwargs):
-    transaction.on_commit(lambda: instance._post_save_event_task(*args, **kwargs))
 
 
 class ImageArchiveFile(BaseImageFile):
